(int)response.StatusCode

+++ rels/Wiki/Web.cs
@@ -23,7 +23,7 @@ namespace rels.Wiki
             {
                 log.OnNext(string.Format("{0} - GetStringAsync - {1}\r\n", DateTime.Now, url));
                 HttpResponseMessage response = await client.GetAsync(url);
-                log.OnNext(string.Format("{0} - {1} - {2}\r\n", DateTime.Now, response.StatusCode, response.ReasonPhrase));
+                log.OnNext(string.Format("{0} - {1} - {2}\r\n", DateTime.Now, (int)response.StatusCode, response.ReasonPhrase));
                 response.EnsureSuccessStatusCode();
                 string responseBody = await response.Content.ReadAsStringAsync();
                 return responseBody;
